search: Move shared zoekt utilities into shared package
Moves some zoekt-related functions that aren't specific to
`graphqlbackend` into a separate package so they can be used from
`searcher` without depending on `graphqlbackend`.
